Encapsulate prompt dismissal state management to react hook
