+++ client/merchant-feedback-prompt/index.tsx
@@ -12,13 +12,13 @@ import {
 	SnackbarList,
 } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
-import { useUserPreferences } from '@woocommerce/data';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import { recordEvent } from 'wcpay/tracks';
+import { useMerchantFeedbackPromptState } from './hooks';
 import './style.scss';
 
 /**
@@ -200,42 +200,24 @@ const MerchantFeedbackPrompt: React.FC< MerchantFeedbackPromptProps > = ( {
 };
 
 /**
- * Extend the user preferences returned from useUserPreferences to include the WooPayments merchant feedback prompt dismissed state.
- * See WC_Payments::add_user_data_fields() in includes/class-wc-payments.php for the PHP implementation.
- */
-interface UserPreferences extends ReturnType< typeof useUserPreferences > {
-	wc_payments_wporg_review_2025_prompt_dismissed: boolean;
-}
-
-/**
  * A wrapper component that conditionally renders the merchant feedback prompt.
  *
  * This is used to ensure the prompt is only rendered if the account is eligible for the campaign and the user has not dismissed the prompt.
  */
 export default function MaybeShowMerchantFeedbackPrompt() {
 	const {
-		updateUserPreferences,
-		...userPrefs
-	} = useUserPreferences() as UserPreferences;
+		isAccountEligible,
+		hasUserDismissedPrompt,
+		setHasUserDismissedPrompt,
+	} = useMerchantFeedbackPromptState();
 
-	const hasUserDismissed =
-		userPrefs?.wc_payments_wporg_review_2025_prompt_dismissed;
-
-	const isAccountEligible =
-		wcpaySettings?.featureFlags?.isMerchantFeedbackPromptDevFlagEnabled &&
-		wcpaySettings?.accountStatus?.campaigns?.wporgReview2025;
-
-	if ( hasUserDismissed || ! isAccountEligible ) {
+	if ( hasUserDismissedPrompt || ! isAccountEligible ) {
 		return null;
 	}
 
 	return (
 		<MerchantFeedbackPrompt
-			dismissPrompt={ () =>
-				updateUserPreferences( {
-					wc_payments_wporg_review_2025_prompt_dismissed: true,
-				} )
-			}
+			dismissPrompt={ () => setHasUserDismissedPrompt( true ) }
 		/>
 	);
 }
